authentication

+++ app/Http/Controllers/TasksController.php
@@ -15,13 +15,22 @@ class TasksController extends Controller
      */
     public function index()
     {
-        // メッセージ一覧を取得
-        $tasks = Task::all();
+        $data = [];
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            // ユーザの投稿の一覧を作成日時の降順で取得
+            // （後のChapterで他ユーザの投稿も取得するように変更しますが、現時点ではこのユーザの投稿のみ取得します）
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        }
         
-        // メッセージ一覧ビューでそれを表示
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+        // Welcomeビューでそれらを表示
+        return view('welcome', $data);
     }
 
     /**
@@ -31,6 +40,7 @@ class TasksController extends Controller
      */
     public function create()
     {
+        
         $task = new Task;
 
         // メッセージ作成ビューを表示
@@ -54,13 +64,20 @@ class TasksController extends Controller
             'content' => 'required|max:255',
         ]);
         
-        
          // メッセージを作成
         $task = new Task;
+        $task->user_id = \Auth::id();
         $task->status = $request->status;    // 追加
         $task->content = $request->content;
         $task->save();
-
+        
+        /**↑は以下でもよい
+        *$request->user()->tasks()->create([
+        *   'status' => $request->status,    // 追加
+        *   'content' => $request->content,
+        *]);
+        */
+        
         // トップページへリダイレクトさせる
         return redirect('/');
     }
@@ -73,13 +90,18 @@ class TasksController extends Controller
      */
     public function show($id)
     {
-        // idの値でメッセージを検索して取得
-        $task = Task::findOrFail($id);
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+        
+            // idの値でメッセージを検索して取得
+            $task = Task::findOrFail($id);
 
-        // メッセージ詳細ビューでそれを表示
-        return view('tasks.show', [
-            'task' => $task,
-        ]);
+            // メッセージ詳細ビューでそれを表示
+            return view('tasks.show', [
+                'task' => $task,
+            ]);
+        }
     }
 
     /**
@@ -90,14 +112,21 @@ class TasksController extends Controller
      */
     public function edit($id)
     {
-        // idの値でメッセージを検索して取得
-        $task = Task::findOrFail($id);
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+        
+            // idの値でメッセージを検索して取得
+            $task = Task::findOrFail($id);
 
-        // メッセージ編集ビューでそれを表示
-        return view('tasks.edit', [
-            'task' => $task,
-        ]);
+            // メッセージ編集ビューでそれを表示
+            return view('tasks.edit', [
+                'task' => $task,
+            ]);
+        }
     }
+        
+        
 
     /**
      * Update the specified resource in storage.
@@ -116,6 +145,7 @@ class TasksController extends Controller
         
         // idの値でメッセージを検索して取得
         $task = Task::findOrFail($id);
+        
         // メッセージを更新
         $task->status = $request->status;    // 追加
         $task->content = $request->content;
@@ -135,8 +165,11 @@ class TasksController extends Controller
     {
         // idの値でメッセージを検索して取得
         $task = Task::findOrFail($id);
-        // メッセージを削除
-        $task->delete();
+        
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は、投稿を削除
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
 
         // トップページへリダイレクトさせる
         return redirect('/');
